Implement equals method in MailData
